[CDF-21686] 🤹‍♂️Structure tests to match package (#643)
* refactor: moved command tests into tests_commands

* tests: split test_loaders in multiple files

* refactor: removed test_templates
