+++ tests/tests_unit/test_cdf_tk/test_commands/test_auth.py
